Address review #2662 - Move the GC-related upgrade code into the upgraders
The upgraders were sharing code with the type system, which the
startTime and endTime changes ended up being a breaking change. In order
to properly isolate this code, it was duplicated and inlined in the
upgraders.
