+++ src/profile-logic/gecko-profile-versioning.js
@@ -11,10 +11,6 @@
  * to the current format.
  */
 
-import {
-  upgradeGCMinorMarker,
-  upgradeGCMajorMarker_Gecko8To9,
-} from './convert-markers';
 import { UniqueStringArray } from '../utils/unique-string-array';
 import { GECKO_PROFILE_VERSION } from '../app-logic/constants';
 
@@ -247,6 +243,73 @@ const _upgraders = {
     convertToVersionEightRecursive(profile);
   },
   [9]: profile => {
+    // Upgrade GC markers
+
+    /*
+     * Upgrade a GCMajor marker in the Gecko profile format.
+     */
+    function upgradeGCMajorMarker_Gecko8To9(marker) {
+      if ('timings' in marker) {
+        if (!('status' in marker.timings)) {
+          /*
+           * This is the old version of the GCMajor marker.
+           */
+
+          const timings = marker.timings;
+
+          timings.status = 'completed';
+
+          /*
+           * The old version had a bug where the slices field could be included
+           * twice with different meanings.  So we attempt to read it as either
+           * the number of slices or a list of slices.
+           */
+          if (Array.isArray(timings.sices)) {
+            timings.slices_list = timings.slices;
+            timings.slices = timings.slices.length;
+          }
+
+          timings.allocated_bytes = timings.allocated * 1024 * 1024;
+        }
+      }
+
+      return marker;
+    }
+
+    function upgradeGCMinorMarker(marker8) {
+      if ('nursery' in marker8) {
+        if ('status' in marker8.nursery) {
+          if (marker8.nursery.status === 'no collection') {
+            marker8.nursery.status = 'nursery empty';
+          }
+          return Object.assign(marker8);
+        }
+        /*
+         * This is the old format for GCMinor, rename some
+         * properties to the more sensible names in the newer
+         * format and set the status.
+         *
+         * Note that we don't delete certain properties such as
+         * promotion_rate, leave them so that anyone opening the
+         * raw json data can still see them in converted profiles.
+         */
+        const marker = Object.assign(marker8, {
+          nursery: Object.assign(marker8.nursery, {
+            status: 'complete',
+            bytes_used: marker8.nursery.nursery_bytes,
+            // cur_capacity cannot be filled in.
+            new_capacity: marker8.nursery.new_nursery_bytes,
+            phase_times: marker8.nursery.timings,
+          }),
+        });
+        delete marker.nursery.nursery_bytes;
+        delete marker.nursery.new_nursery_bytes;
+        delete marker.nursery.timings;
+        return marker;
+      }
+      return marker8;
+    }
+
     function convertToVersionNineRecursive(p) {
       for (const thread of p.threads) {
         //const stringTable = new UniqueStringArray(thread.stringTable);
